+++ src/ucvm/ucvm.c
@@ -639,10 +639,13 @@ int ucvm_assoc_ifunc(const char *mlabel, const char *ilabel)
   if (strcmp(ilabel, UCVM_IFUNC_ELY) == 0) {
     ucvm_strcpy(ifunc.label, UCVM_IFUNC_ELY, UCVM_MAX_LABEL_LEN);
     ifunc.interp = ucvm_interp_ely;
+  } else if (strcmp(ilabel, UCVM_IFUNC_SO2) == 0) {
+    ucvm_strcpy(ifunc.label, UCVM_IFUNC_SO2, UCVM_MAX_LABEL_LEN);
+    ifunc.interp = ucvm_interp_so2;
   } else if (strcmp(ilabel, UCVM_IFUNC_TAPER) == 0) {
     ucvm_strcpy(ifunc.label, UCVM_IFUNC_TAPER, UCVM_MAX_LABEL_LEN);
     ifunc.interp = ucvm_interp_taper;
-    // need to reset the zrange
+    // need to force reset the zrange
     ucvm_setparam(UCVM_MODEL_PARAM_IFUNC_ZRANGE, 0.0, 700.0);
   } else if (strcmp(ilabel, UCVM_IFUNC_CRUST) == 0) {
     ucvm_strcpy(ifunc.label, UCVM_IFUNC_CRUST, UCVM_MAX_LABEL_LEN);
@@ -1355,6 +1358,13 @@ int ucvm_get_resources(ucvm_resource_t *res, int *len)
     return(UCVM_CODE_ERROR);
   }
 #endif
+#ifdef _UCVM_ENABLE_SHAKEOUT2
+  if (ucvm_save_resource(UCVM_RESOURCE_MODEL, UCVM_MODEL_CRUSTAL,
+                     UCVM_MODEL_SHAKEOUT2, "", res, numinst++, *len)
+      != UCVM_CODE_SUCCESS) {
+    return(UCVM_CODE_ERROR);
+  }
+#endif
 #ifdef _UCVM_ENABLE_MUSCALNC
   if (ucvm_save_resource(UCVM_RESOURCE_MODEL, UCVM_MODEL_CRUSTAL,
                      UCVM_MODEL_MUSCALNC, "", res, numinst++, *len)
@@ -1503,6 +1513,12 @@ int ucvm_get_resources(ucvm_resource_t *res, int *len)
   /* Get installed ifuncs */
   if (ucvm_save_resource(UCVM_RESOURCE_IFUNC, UCVM_MODEL_CRUSTAL,
 		     UCVM_IFUNC_ELY, "", res, numinst++, *len) 
+      != UCVM_CODE_SUCCESS) {
+    return(UCVM_CODE_ERROR);
+  }
+
+  if (ucvm_save_resource(UCVM_RESOURCE_IFUNC, UCVM_MODEL_CRUSTAL,
+                     UCVM_IFUNC_SO2, "", res, numinst++, *len)
       != UCVM_CODE_SUCCESS) {
     return(UCVM_CODE_ERROR);
   }