+++ src/ucvm/ucvm_dtypes.h
@@ -86,6 +86,7 @@
 #define UCVM_MODEL_SJFZ "sjfz"
 #define UCVM_MODEL_CS248 "cs248"
 
+#define UCVM_MODEL_SHAKEOUT2 "shakeout2"
 #define UCVM_MODEL_MUSCALNC "muscalnc"
 #define UCVM_MODEL_MUSCALTDB "muscaltdb"
 
@@ -142,6 +143,7 @@
 #define UCVM_IFUNC_GTL "gtl"
 #define UCVM_IFUNC_LINEAR "linear"
 #define UCVM_IFUNC_ELY "ely"
+#define UCVM_IFUNC_SO2 "so2"
 #define UCVM_IFUNC_TAPER "taper"