+++ src/ucvm/ucvm_interp.h
@@ -7,6 +7,10 @@
 int ucvm_interp_ely(double zmin, double zmax, ucvm_ctype_t cmode,
 		    ucvm_point_t *pnt, ucvm_data_t *data);
 
+/* ShakeOut2 interpolation method */
+int ucvm_interp_so2(double zmin, double zmax, ucvm_ctype_t cmode,
+                    ucvm_point_t *pnt, ucvm_data_t *data);
+
 /* Taper interpolation method */
 int ucvm_interp_taper(double zmin, double zmax, ucvm_ctype_t cmode,
                     ucvm_point_t *pnt, ucvm_data_t *data);