+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2026-10-XX Mei-Hui Su <mei@usc.edu>
+* TARGET release UCVM_2610 
+
+## 2026-9-30 Mei-Hui Su <mei@usc.edu>
+* Added SHAKEOUT2 (NetCDF)
+
 ## 2026-08-13 Mei-Hui Su <mei@usc.edu>
 * Tag UCVM CVWS2026.8